chore: 상태관리하는 폴더명 변경
